+++ Core/Inc/dribbler_board/Util/gpio_util.h
@@ -8,18 +8,6 @@
 #ifndef UTILS_GPIO_UTIL_H_
 #define UTILS_GPIO_UTIL_H_
 
-#include "stdbool.h"
-#include "stm32f3xx_hal.h"
-
-#define LOW 0
-#define HIGH 1
-
-// abstract a GPIO pin combination to a struct
-typedef struct GPIO_Pin{
-	GPIO_TypeDef * PORT;
-	uint16_t PIN;
-} GPIO_Pin;
-
 /////////////////////////////////// LIST OF KNOWN GPIO PINS
 
 // ball sensor
@@ -28,25 +16,5 @@ extern GPIO_Pin IR_LED_pin;
 // LED
 extern GPIO_Pin LED1;
 extern GPIO_Pin LED2;
-
-/////////////////////////////////////////////// GPIO UTILITY FUNCTIONS
-
-// Set a GPIO Pin
-inline void set_Pin(GPIO_Pin p, bool value)
-{
-	HAL_GPIO_WritePin(p.PORT, p.PIN, value);
-}
-
-// Read a GPIO Pin
-inline GPIO_PinState read_Pin(GPIO_Pin p)
-{
-	return HAL_GPIO_ReadPin(p.PORT, p.PIN);
-}
-
-// Toggle a GPIO Pin
-inline void toggle_Pin(GPIO_Pin p)
-{
-	HAL_GPIO_TogglePin(p.PORT, p.PIN);
-}
 
 #endif /* UTILS_GPIO_UTIL_H_ */
