+++ Core/Inc/general/Util/gpio_util_functions.h
@@ -1,0 +1,39 @@
+#ifndef UTILS_GPIO_UTIL_FUNCTIONS_H
+#define UTILS_GPIO_UTIL_FUNCTIONS_H
+
+#include "stdbool.h"
+#include "main.h" //circular include because of different microcontrollers, platformio handles circular includes
+
+#define LOW 0
+#define HIGH 1
+
+#define DEFINE_PIN(label,pin_name) GPIO_Pin label = {pin_name##_GPIO_Port, pin_name##_Pin};
+
+// abstract a GPIO pin combination to a struct
+typedef struct GPIO_Pin{
+	GPIO_TypeDef * PORT;
+	uint16_t PIN;
+} GPIO_Pin;
+
+/////////////////////////////////////////////// GPIO UTILITY FUNCTIONS
+
+// Set a GPIO Pin
+inline void set_Pin(GPIO_Pin p, bool value)
+{
+	HAL_GPIO_WritePin(p.PORT, p.PIN, value);
+}
+
+// Read a GPIO Pin
+inline GPIO_PinState read_Pin(GPIO_Pin p)
+{
+	return HAL_GPIO_ReadPin(p.PORT, p.PIN);
+}
+
+// Toggle a GPIO Pin
+inline void toggle_Pin(GPIO_Pin p)
+{
+	HAL_GPIO_TogglePin(p.PORT, p.PIN);
+}
+
+
+#endif /* UTILS_GPIO_UTIL_FUNCTIONS_H */
