+++ Core/Src/kicker_chipper_board/Util/gpio_util.c
@@ -8,7 +8,6 @@
 #include "gpio_util.h"
 #include "main.h"
 
-#define DEFINE_PIN(label,pin_name) GPIO_Pin label = {pin_name##_GPIO_Port, pin_name##_Pin};
 // List known GPIO pins
 
 // Kick/Chip
